added winston to routes and modularized lambda fetch
