+++ mail-aliases/src/app/api/route.js
@@ -1,4 +1,5 @@
 import { logger } from '@/serverComponents/logger';
+import { sendApiRequest } from '@/serverComponents/executeLambda';
 
 // Signing code based on:
 //   https://arpadt.com/articles/signing-requests-with-aws-sdk
@@ -78,67 +79,3 @@ async function searchDatabase(searchString) {
   return responseJson;
 }
 
-
-async function sendApiRequest(requestMethod = 'GET', endpoint, payload = {}) {
-  const baseUrl = process.env.LAMBDA_URL;
-  const apiUrl = new URL(endpoint, baseUrl);
-
-  // Prepare to sign the request
-  const sigv4 = new SignatureV4({
-    service: 'lambda',
-    region: 'us-east-1',
-    credentials: {
-      accessKeyId: process.env.AWS_ACCESS_KEY,
-      secretAccessKey: process.env.AWS_SECRET_KEY,
-    },
-    sha256: Sha256,
-  });
-
-  // Build singing Payload
-  const signPayload = {
-    method: requestMethod,
-    hostname: apiUrl.host,
-    path: apiUrl.pathname,
-    protocol: apiUrl.protocol,
-    headers: {
-      'Content-Type': 'application/json',
-      host: apiUrl.hostname,
-    },
-  };
-
-  // Check query endpoint for URL Parameters
-  console.log('route:searchParams -- ' + apiUrl.searchParams.toString());
-  if (apiUrl.searchParams.toString().length > 0) {
-    const params = apiUrl.searchParams.toString().split('&');
-    const canonicalQueryObject = {};
-
-    params.forEach(element => {
-      const [p, v] = element.split('=');
-      canonicalQueryObject[decodeURIComponent(p)] = decodeURIComponent(v);
-    });
-
-    signPayload.query = canonicalQueryObject;
-  }
-
-  // Check if there is a payload
-  console.log('route:payload -- ' + JSON.stringify(payload));
-  if (Object.keys(payload).length > 0) signPayload.body = JSON.stringify(payload);
-
-  // Get Signature
-  const signed = await sigv4.sign(signPayload);
-  console.log('route:signedRequest -- ' + JSON.stringify(signed));
-
-  // fetch
-  const fetchParams = {
-    method: requestMethod,
-    headers: signed.headers,
-  };
-
-  // If there is a payload, attach it to the body.
-  if (Object.keys(payload).length > 0) fetchParams.body = JSON.stringify(payload);
-  const fetchResults = await fetch(apiUrl.href, fetchParams);
-
-  const data = await fetchResults.json();
-  console.log('route:sendApiRequest -- ' + JSON.stringify(data));
-  return data;
-}
